# src/main/resources/static/swagger-ui/openapi3.yaml
openapi: 3.0.1
info:
  title: 801base API docs
  description: 801base의 API 문서입니다.
  version: 0.0.1
servers:
- url: https://api2.801base.com
...
paths:
...
  /admin/orders/{orderId}:
    get:
      tags:
      - Admin-Order
      summary: 주문 상세 조회
      operationId: 관리자_주문상세_조회_성공
      parameters:
      - name: orderId
        in: path
        description: 주문 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-orders-orderId1238496271"
              examples:
                관리자_주문상세_조회_성공:
                  value: |-
                    {
                      "data" : {
                        "orderNumber" : "ORD123",
                        "status" : "배송중",
                        "trackingNumber" : "234980234",
                        "createdAt" : "2025-06-10T12:00",
                        "paymentMethod" : "MOCK",
                        "address" : "서울시 관악구",
                        "recipientName" : "홍길동",
                        "recipientPhone" : "0101234",
                        "customerName" : "커피매니아",
                        "customerEmail" : "커피매니아@mail.com",
                        "items" : [ {
                          "productName" : "스타벅스 캡슐",
                          "quantity" : 1,
                          "price" : 1000,
                          "total" : 1000,
                          "thumbnail" : "http://asdf.com?thumb.jpg"
                        } ],
                        "subtotal" : 1000,
                        "total" : 1000
                      },
                      "error" : null
                    }
...
  /admin/orders/{orderId}/cancel:
    delete:
      tags:
      - Admin-Order
      summary: 관리자 주문 취소
      operationId: 관리자_주문취소_성공
      parameters:
      - name: orderId
        in: path
        description: 주문 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-orders-orderId-cancel486549215"
              examples:
                관리자_주문취소_성공:
                  value: |-
                    {
                      "data" : null,
                      "error" : null
                    }
...
  /orders:
    get:
      tags:
      - Order
      summary: 주문 목록을 조회합니다
      operationId: 주문_목록_조회_성공
      parameters:
      - name: monthRange
        in: query
        description: "주문 월(3,6,12개월)"
        required: false
        schema:
          type: string
      - name: status
        in: query
        description: 주문상태
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: "페이지 번호 (기본값: 1)"
        required: false
        schema:
          type: string
      - name: Authorization
        in: header
        description: 인증 토큰
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orders-1230386813"
              examples:
                주문_목록_조회_성공:
                  value: |-
                    {
                      "data" : {
                        "content" : [ {
                          "orderNumber" : "ORD20250609123456789",
                          "orderName" : "스페셜 리버즈 외 3건",
                          "mainProductThumbnail" : "https://example.com/thumbnail.jpg",
                          "orderStatus" : "DELIVERED",
                          "finalTotalPrice" : 13000,
                          "orderedAt" : "2025-06-08T12:34",
                          "trackingNumber" : "4231908234098",
                          "cancellable" : true,
                          "refundable" : false
                        } ],
                        "page" : 2,
                        "size" : 10,
                        "totalPages" : 2,
                        "totalElements" : 11
                      },
                      "error" : null
                    }
    post:
      tags:
      - Order
      summary: 주문을 생성합니다
      operationId: 주문_생성_성공
      parameters:
      - name: Authorization
        in: header
        description: 인증 토큰
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/orders-1934294331"
            examples:
              주문_생성_성공:
                value: |-
                  {
                    "cartItemIds" : [ 1, 2, 3 ],
                    "shippingInfo" : {
                      "recipientName" : "홍길동",
                      "recipientPhone" : "010-1234-1234",
                      "zipCode" : "12345",
                      "address1" : "서울특별시 관악구",
                      "address2" : "서울대입구역 6번출구",
                      "deliveryMessage" : "문앞에 놔주세요"
                    },
                    "paymentMethod" : "TOSS_PAY"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orders-841088526"
              examples:
                주문_생성_성공:
                  value: |-
                    {
                      "data" : {
                        "orderNumber" : "ORD20250609123456789"
                      },
                      "error" : null
                    }
  /orders/prepare:
    get:
      tags:
      - Order
      summary: 주문서 생성
      operationId: 주문서_생성_성공
      parameters:
      - name: cartItemIds
        in: query
        description: 장바구니 아이템 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orders-prepare-1852172854"
              examples:
                주문서_생성_성공:
                  value: |-
                    {
                      "data" : {
                        "cartItemIds" : [ 1 ],
                        "itemsSubtotal" : 10000,
                        "shippingFee" : 3000,
                        "finalTotalPrice" : 13000,
                        "items" : [ {
                          "cartItemId" : 1,
                          "productId" : 1,
                          "name" : "상품A",
                          "thumbnail" : "http://localhost:8080/api/v1/product/1/thumbnail",
                          "unitPrice" : 1000,
                          "quantity" : 10,
                          "itemSubtotal" : 10000
                        } ],
                        "shippingInfo" : {
                          "recipientName" : "홍길동",
                          "recipientPhone" : "010-1234-1234",
                          "zipCode" : "12345",
                          "addressId" : 1,
                          "address1" : "서울특별시 관악구",
                          "address2" : "서울대입구역 6번출구"
                        },
                        "paymentMethod" : [ {
                          "code" : "MOCK",
                          "label" : "테스트"
                        } ]
                      },
                      "error" : null
                    }
  /orders/status:
    get:
      tags:
      - Order
      summary: 주문 상태 코드 목록
      operationId: 주문상태_코드목록_조회_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orders-status436844904"
              examples:
                주문상태_코드목록_조회_성공:
                  value: |-
                    {
                      "data" : [ {
                        "code" : "WAITING_FOR_PAYMENT",
                        "label" : "결제 대기중"
                      } ],
                      "error" : null
                    }
  /orders/{orderNumber}:
    get:
      tags:
      - Order
      summary: 주문 상세 정보를 조회합니다
      operationId: 주문_상세_조회_성공
      parameters:
      - name: orderNumber
        in: path
        description: 주문 번호
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: 인증 토큰
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orders-orderNumber-1027130959"
              examples:
                주문_상세_조회_성공:
                  value: |-
                    {
                      "data" : {
                        "orderNumber" : "ORD20250609123456789",
                        "orderName" : "홍길동님의 주문",
                        "orderStatus" : "배송 준비중",
                        "trackingNumber" : "234980234",
                        "paymentNumber" : "PAY1231414124",
                        "paymentMethod" : "토스페이",
                        "itemsSubTotal" : 10000,
                        "shippingFee" : 3000,
                        "finalTotalPrice" : 13000,
                        "items" : [ {
                          "orderItemId" : 1,
                          "productSnapshotId" : 1,
                          "name" : "상품명",
                          "thumbnail" : "https://example.com/thumbnail.jpg",
                          "unitPrice" : 1000,
                          "quantity" : 10,
                          "itemSubTotal" : 10000
                        } ],
                        "shippingInfo" : {
                          "recipientName" : "홍길동",
                          "recipientPhone" : "010-1234-1234",
                          "zipCode" : "08123",
                          "address1" : "서울특별시 관악구",
                          "address2" : "1000003동 123호",
                          "deliveryMessage" : "문앞에 놔주세요."
                        },
                        "orderedAt" : "2025-06-08T12:34",
                        "paidAt" : "2025-06-08T12:34",
                        "cancellable" : true,
                        "cancelRequested" : false,
                        "cancelledAt" : null,
                        "refundable" : false,
                        "refundRequested" : false,
                        "refundRequestedAt" : null,
                        "refunded" : false,
                        "refundedAt" : null,
                        "reviewable" : true,
                        "reviewWritten" : false
                      },
                      "error" : null
                    }
  /orders/{orderNumber}/cancel:
    post:
      tags:
      - Order
      summary: 주문을 취소합니다
      operationId: 주문_취소_성공
      parameters:
      - name: orderNumber
        in: path
        description: 주문 번호
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: 인증 토큰
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/admin-orders-orderId-cancel486549215"
            examples:
              주문_취소_성공:
                value: "{ }"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-reviews-reviewId-reply1686665306"
              examples:
                주문_취소_성공:
                  value: |-
                    {
                      "data" : {
                        "message" : "OK"
                      },
                      "error" : null
                    }
  /payments:
    post:
      tags:
      - Payment
      summary: 결제 처리 한다.
      description: |2-
                        ORD-001: 주문 내역을 찾을 수 없습니다.
                        PAY-001: PG사 결제정보를 찾을 수 업습니다.
                        PAY-002: 결제 정보를 찾을 수 없습니다.
                        PAY-003: 이미 결제처리한 주문입니다.
                        PAY-004: 이미 결제되었습니다.
                        PAY-005: 결제금액이 일치하지 않습니다.
      operationId: 결제처리_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/payments808279509"
            examples:
              결제처리_성공:
                value: |-
                  {
                    "orderNumber" : "ORD123123123",
                    "transactionId" : "12312312-123"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/payments-2018243494"
              examples:
                결제처리_성공:
                  value: |-
                    {
                      "data" : {
                        "paymentNumber" : "PAY00001"
                      },
                      "error" : null
                    }
  /payments/cancel:
    post:
      tags:
      - Payment
      summary: 결제 취소 한다.
      description: |2-
                        ORD-001: 주문 내역을 찾을 수 없습니다.
                        PAY-001: PG사 결제정보를 찾을 수 업습니다.
                        PAY-002: 결제 정보를 찾을 수 없습니다.
                        PAY-006: 결제 대기중, 결제 완료 상태인 주문만 취소할 수 있습니다.
                        PAY-007: 다른 사람의 결제를 취소할 수 없습니다.
                        PAY-008: PG사 결제 아이디가 누락되었습니다.
      operationId: 결제취소_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/payments-cancel-919447853"
            examples:
              결제취소_성공:
                value: |-
                  {
                    "orderNumber" : "ORD123123123"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-reviews-reviewId-reply1686665306"
              examples:
                결제취소_성공:
                  value: |-
                    {
                      "data" : {
                        "message" : "OK"
                      },
                      "error" : null
                    }
  /payments/refund:
    post:
      tags:
      - Payment
      summary: 결제 환불요청 한다.
      description: |2-
                        ORD-001: 주문 내역을 찾을 수 없습니다.
                        PAY-009: 배송중, 배송 완료 상태인 주문만 환불할 수 있습니다.
                        PAY-0010: 다른 사람의 결제를 환불할 수 없습니다.
      operationId: 결제환불요청_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/payments-cancel-919447853"
            examples:
              결제환불요청_성공:
                value: |-
                  {
                    "orderNumber" : "ORD123123123"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-reviews-reviewId-reply1686665306"
              examples:
                결제환불요청_성공:
                  value: |-
                    {
                      "data" : {
                        "message" : "OK"
                      },
                      "error" : null
                    }
...
components:
  schemas:
...
    orders-1230386813:
      type: object
      properties:
        data:
          required:
          - page
          - size
          - totalElements
          - totalPages
          type: object
          properties:
            size:
              type: number
              description: 페이지 사이즈(기본값 10)
            totalPages:
              type: number
              description: 전체 페이지 수
            page:
              type: number
              description: 현재 페이지 (기본값 1)
            content:
              type: array
              items:
                required:
                - cancellable
                - finalTotalPrice
                - mainProductThumbnail
                - orderName
                - orderNumber
                - orderStatus
                - orderedAt
                - refundable
                type: object
                properties:
                  orderedAt:
                    type: string
                    description: 주문 날짜
                  mainProductThumbnail:
                    type: string
                    description: 대표 상품 썸네일
                  orderNumber:
                    type: string
                    description: 주문 번호
                  finalTotalPrice:
                    type: number
                    description: 최종 결제 금액
                  orderStatus:
                    type: string
                    description: 주문 상태
                  refundable:
                    type: boolean
                    description: 환불 가능 여부
                  cancellable:
                    type: boolean
                    description: 취소 가능 여부
                  trackingNumber:
                    type: string
                    description: 송장번호
                    nullable: true
                  orderName:
                    type: string
                    description: 주문명
            totalElements:
              type: number
              description: 총 수
    orders-841088526:
      type: object
      properties:
        data:
          required:
          - orderNumber
          type: object
          properties:
            orderNumber:
              type: string
              description: 주문 번호
...
    orders-prepare-1852172854:
      type: object
      properties:
        data:
          required:
          - cartItemIds
          - finalTotalPrice
          - itemsSubtotal
          - shippingFee
          type: object
          properties:
            itemsSubtotal:
              type: number
              description: 상품 금액 합계
            shippingFee:
              type: number
              description: 배송비
            cartItemIds:
              type: array
              description: 장바구니 아이템 아이디
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            finalTotalPrice:
              type: number
              description: 최종 결제 금액
            shippingInfo:
              required:
              - address1
              - addressId
              - recipientName
              - recipientPhone
              - zipCode
              type: object
              properties:
                zipCode:
                  type: string
                  description: 우편번호
                recipientPhone:
                  type: string
                  description: 수령인 전화번호
                address2:
                  type: string
                  description: 상세주소
                  nullable: true
                address1:
                  type: string
                  description: 주소
                recipientName:
                  type: string
                  description: 수령인 이름
                addressId:
                  type: number
                  description: 배송지 아이디
            paymentMethod:
              type: array
              items:
                required:
                - code
                - label
                type: object
                properties:
                  code:
                    type: string
                    description: 결제 수단 코드
                  label:
                    type: string
                    description: 결제 수단 이름
            items:
              type: array
              items:
                required:
                - cartItemId
                - itemSubtotal
                - name
                - productId
                - quantity
                - thumbnail
                - unitPrice
                type: object
                properties:
                  itemSubtotal:
                    type: number
                    description: 상품 소계
                  unitPrice:
                    type: number
                    description: 상품 단가
                  thumbnail:
                    type: string
                    description: 상품 썸네일 URL
                  quantity:
                    type: number
                    description: 수량
                  productId:
                    type: number
                    description: 상품 ID
                  name:
                    type: string
                    description: 상품명
                  cartItemId:
                    type: number
                    description: 장바구니 ID
...
    admin-reviews-reviewId-reply1686665306:
      type: object
      properties:
        data:
          required:
          - message
          type: object
          properties:
            message:
              type: string
              description: 응답 메시지
...
    admin-orders-orderId1238496271:
      type: object
      properties:
        data:
          required:
          - address
          - createdAt
          - customerEmail
          - customerName
          - orderNumber
          - paymentMethod
          - recipientName
          - recipientPhone
          - status
          - subtotal
          - total
          type: object
          properties:
            recipientPhone:
              type: string
              description: 주문번호
            address:
              type: string
              description: 주문번호
            orderNumber:
              type: string
              description: 주문번호
            customerName:
              type: string
              description: 주문번호
            createdAt:
              type: string
              description: 주문번호
            total:
              type: number
              description: 주문번호
            subtotal:
              type: number
              description: 주문번호
            customerEmail:
              type: string
              description: 주문번호
            recipientName:
              type: string
              description: 주문번호
            paymentMethod:
              type: string
              description: 주문번호
            items:
              type: array
              items:
                required:
                - price
                - productName
                - quantity
                - total
                type: object
                properties:
                  thumbnail:
                    type: string
                    description: 주문번호
                    nullable: true
                  total:
                    type: number
                    description: 주문번호
                  quantity:
                    type: number
                    description: 주문번호
                  price:
                    type: number
                    description: 주문번호
                  productName:
                    type: string
                    description: 주문번호
            trackingNumber:
              type: string
              description: 송장번호
              nullable: true
            status:
              type: string
              description: 주문번호
...
    admin-orders-orderId-cancel486549215:
      type: object
...
    orders-status436844904:
      type: object
      properties:
        data:
          type: array
          items:
            required:
            - code
            - label
            type: object
            properties:
              code:
                type: string
                description: 주문상태 코드
              label:
                type: string
                description: 주문상태 라벨
...
    payments-cancel-919447853:
      required:
      - orderNumber
      type: object
      properties:
        orderNumber:
          type: string
          description: 주분번호
...
    payments-2018243494:
      type: object
      properties:
        data:
          required:
          - paymentNumber
          type: object
          properties:
            paymentNumber:
              type: string
              description: 결제번호
    orders-1934294331:
      required:
      - cartItemIds
      - paymentMethod
      - shippingInfo
      type: object
      properties:
        cartItemIds:
          type: array
          description: 장바구니 아이템 ID 목록
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        shippingInfo:
          type: object
          description: 배송지 정보
        paymentMethod:
          type: string
          description: 결제 수단
...
    orders-orderNumber-1027130959:
      type: object
      properties:
        data:
          required:
          - cancelRequested
          - cancellable
          - finalTotalPrice
          - itemsSubTotal
          - orderName
          - orderNumber
          - orderStatus
          - orderedAt
          - paidAt
          - paymentMethod
          - paymentNumber
          - refundRequested
          - refundable
          - refunded
          - reviewWritten
          - reviewable
          - shippingFee
          type: object
          properties:
            orderNumber:
              type: string
              description: 주문 번호
            shippingInfo:
              required:
              - address1
              - address2
              - deliveryMessage
              - recipientName
              - recipientPhone
              - zipCode
              type: object
              properties:
                zipCode:
                  type: string
                  description: 우편번호
                recipientPhone:
                  type: string
                  description: 수령인 전화번호
                address2:
                  type: string
                  description: 상세주소
                address1:
                  type: string
                  description: 주소
                recipientName:
                  type: string
                  description: 수령인 이름
                deliveryMessage:
                  type: string
                  description: 배송 메시지
            itemsSubTotal:
              type: number
              description: 상품 금액 합계
            refundRequested:
              type: boolean
              description: 환불 요청 여부
            orderStatus:
              type: string
              description: 주문 상태
            cancellable:
              type: boolean
              description: 취소 가능 여부
            orderedAt:
              type: string
              description: 주문 일시
            reviewable:
              type: boolean
              description: 리뷰 작성 가능 여부
            shippingFee:
              type: number
              description: 배송비
            finalTotalPrice:
              type: number
              description: 최종 결제 금액
            reviewWritten:
              type: boolean
              description: 리뷰 작성 여부
            paidAt:
              type: string
              description: 결제 일시
            paymentMethod:
              type: string
              description: 결제 수단
            refunded:
              type: boolean
              description: 환불 완료 여부
            refundable:
              type: boolean
              description: 환불 가능 여부
            cancelRequested:
              type: boolean
              description: 취소 요청 여부
            items:
              type: array
              items:
                required:
                - itemSubTotal
                - name
                - orderItemId
                - productSnapshotId
                - quantity
                - thumbnail
                - unitPrice
                type: object
                properties:
                  unitPrice:
                    type: number
                    description: 상품 단가
                  thumbnail:
                    type: string
                    description: 상품 썸네일 URL
                  itemSubTotal:
                    type: number
                    description: 상품 소계
                  quantity:
                    type: number
                    description: 수량
                  orderItemId:
                    type: number
                    description: 주문 아이템 ID
                  name:
                    type: string
                    description: 상품명
                  productSnapshotId:
                    type: number
                    description: 상품 스냅샷 ID
            trackingNumber:
              type: string
              description: 송장번호
              nullable: true
            paymentNumber:
              type: string
              description: 결제 번호
            orderName:
              type: string
              description: 주문명
    payments808279509:
      required:
      - orderNumber
      - transactionId
      type: object
      properties:
        orderNumber:
          type: string
          description: 주분번호
        transactionId:
          type: string
          description: PG사 결제 아이디
...